refactor and log improvements
