+++ .gitignore
@@ -5,6 +5,12 @@
 *.so
 *.dylib
 
+# vscode
+.vscode
+
+# direnv
+.envrc
+
 # Test binary, built with `go test -c`
 *.test
