+++ pkg/chain/chain.go
@@ -2,6 +2,7 @@ package chain
 
 import (
 	"context"
+	"fmt"
 
 	"github.com/cosmos/relayer/v2/relayer"
 	"github.com/cosmos/relayer/v2/relayer/chains/cosmos"
@@ -53,4 +54,20 @@ func PrepChain(info Info) (*relayer.Chain, error) {
 	}
 
 	return chain, nil
+}
+
+func ValidateChainInfo(info Info) error {
+	if info.ChainID == "" {
+		return fmt.Errorf("missing chain ID: %v", info)
+	}
+
+	if info.RPCAddr == "" {
+		return fmt.Errorf("missing RPC address: %v", info)
+	}
+
+	if info.ClientID == "" {
+		return fmt.Errorf("missing client ID: %v", info)
+	}
+
+	return nil
 }
